Search results and viewer work again
