+++ src/lib/api/flatpages.ts
@@ -1,7 +1,9 @@
 import { BASE_API_URL } from "@/config/config";
 import type { Flatpage } from "./types";
 
-export async function getTipOfDay(fetch = globalThis.fetch): Promise<Flatpage> {
+export async function getTipOfDay(
+  fetch = globalThis.fetch,
+): Promise<Flatpage | undefined> {
   const endpoint = new URL("flatpages/tipofday/", BASE_API_URL);
   try {
     const resp = await fetch(endpoint, { credentials: "include" });
